\name{CycleRSI}
\alias{CycleRSI}
\title{Cycle Analytics For Traders RSI Oscillator}
\usage{
  CycleRSI(x, n = 20)
}
\arguments{
  \item{x}{a time series}

  \item{n}{a lookback period}
}
\value{
  an RSI between 0 and 100
}
\description{
  Computes the modified RSI from Chapter 7-4 of Cycle
  Analytics for Traders by Dr. John Ehlers
}

