\name{CycleStoch}
\alias{CycleStoch}
\title{Cycle Analytics For Traders Stochastic Oscillator}
\usage{
  CycleStoch(x, n = 20)
}
\arguments{
  \item{x}{a time series}

  \item{n}{a lookback period}
}
\value{
  a stochastic oscillator between 0 and 1
}
\description{
  Computes the modified stochastic oscillator from Chapter
  7-3 of Cycle Analytics For Traders by Dr. John Ehlers
}

