\name{iFish}
\alias{iFish}
\title{Inverse Fisher Transform}
\usage{
  iFish(x)
}
\arguments{
  \item{x}{an xts or stream of numeric values}
}
\value{
  an object the same type as above
}
\description{
  Inverse Fisher Transform
}
\references{
  \cr
  \url{http://www.mesasoftware.com/Papers/THE\%20INVERSE\%20FISHER\%20TRANSFORM.pdf}\cr
}

