\name{RVI}
\alias{RVI}
\title{Ehlers Relative Vigor Index}
\usage{
  RVI(OHLC, n = 10)
}
\arguments{
  \item{OHLC}{an OHLC xts object}

  \item{n}{number of periods to use}
}
\value{
  an XTS object
}
\description{
  Ehlers Relative Vigor Index
}
\references{
  the following site was used to code this indicator: \cr
  \url{http://www.davenewberg.com/Trading/TS_Code/Ehlers_Indicators/Relative_Vigor_Index.html}\cr
}

