\name{SWI}
\alias{SWI}
\title{Ehlers Sinewave Indicator (v. 2013)}
\usage{
  SWI(x, duration = 40, ...)
}
\arguments{
  \item{x}{a price series}

  \item{duration}{a lookback period}
}
\value{
  The Sinewave Indicator (Oscillator type)
}
\description{
  Ehlers Sinewave Indicator (v. 2013)
}
\references{
  Cycle Analytics For Traders, Chapter 12
}

